requirements.txt
pyTelegramBotAPI
numpy
Pillow==9.0.0